Create tests.py
My name is Colette, so the first and last letter of my name (C + E = 3 + 5) give the number  8. The 8th Fibonacci number is 21.

My surname is Gallagher

The first letter G is number 71

The last letter r is number 114

Fibonacci number 185 is 205697230343233228174223751303346572685

Ord() defines the position of something in a series, such as ‘first’, ‘second’, or ‘third’.  It is useful for converting a single character to the characters corresponding ASCII value.  For example ord (‘a’) returns the integer 97