Update collatz.py
*The Collatz Conjecture
The Collatz conjecture is a conjecture in mathematics that concerns a sequence defined as follows: start with any positive integer n. Then each term is obtained from the previous term as follows: if the previous term is even, the next term is one half the previous term. Otherwise, the next term is 3 times the previous term plus 1. The conjecture is that no matter what value of n, the sequence will always reach 1. (https://en.wikipedia.org/wiki/Collatz_conjecture)

**Exercise 3
This exercise contains a Python script which starts with an integer.  The Collatz Function is then repeatedly applied (divide by 2 if even, multiply by three and add one if odd)

The starting value has been given as 17.  The current value of the integer is printed to the screen at each iteration.